Added Description, JointsPerTier to Rack and Index to Pipe (updated DB and test data), and made all remaining DB functions async.

+++ Inventory-Models/Dto/TallyQuery/DtoTally_WithPipeAndCustomer.cs
@@ -38,8 +38,8 @@ namespace Inventory_Dto.Dto
         public string? CarrierName {  get; set; }
 
         //Calculated Fields
-        public float WeightInKg { get; set; }
-        public float WeightInLbs { get; set; }
+        public Decimal WeightInKg { get; set; }
+        public Decimal WeightInLbs { get; set; }
 
     }
 }
